requirements.txt
flask
pymongo
python-dotenv
openai
pdfminer.six
langchain
pymupdf
flask-cors>=4.0.0
